Move DependenciesSlice to dependencyutils
